requirements.txt
streamlit>=1.47
python-docx>=1.1
openai>=1.30
python-dotenv>=1.0
tqdm>=4.66
